+++ smartgraph.properties
@@ -24,5 +24,6 @@ edge.arrow = true
 #   -- particular problem, knowing that not all will achieve 
 #   -- a stable state
 layout.repulsive-force = 25000
+layout.repulsive-scale = 1
 layout.attraction-force = 30
 layout.attraction-scale = 10
